hotov cities controller a kus measurement, TODO: full measurement controller, testy

+++ ppj_WeatherProject/src/main/java/cz/tul/model/City.java
@@ -88,7 +88,7 @@ public class City {
         return Objects.hash(getId(), getCityName(), getState());
     }
 
-    public long getCityId() {
+    public int getCityId() {
         return cityId;
     }
